refactor: rewrite pagination
