--- src/components/CustomPagination/CustomPagination.module.css
@@ -1,23 +1,0 @@
-.root {
-  display: flex;
-}
-
-.control {
-  font-size: var(--mantine-font-size-sm);
-
-  &:first-child:not(:disabled),
-  &:last-child:not(:disabled) {
-    color: var(--mantine-color-grey-6);
-  }
-
-  &:first-child:disabled,
-  &:last-child:disabled {
-    color: var(--mantine-color-grey-3);
-    opacity: 1;
-    border-color: var(--mantine-color-grey-3);
-  }
-}
-
-.dots {
-  display: none;
-}
